PacketAddParticle PacketDeleteParticle PacketPlaySound PacketStopSound PacketControl

+++ include/network/packet/PacketAddEntity.hh
@@ -18,33 +18,35 @@ namespace gauntlet {
                 unsigned int entityId;
                 unsigned int textureId;
                 unsigned int meshId;
-	      int x;
-                int y;
+                double x;
+                double y;
                 short angle;
             }__attribute__((packed));
 
         public:
             PacketAddEntity(const s_socketData &data);
-            PacketAddEntity(unsigned int entityId, unsigned int textureId, unsigned int meshId, int x, int y, short angle);
+            PacketAddEntity(unsigned int entityId, unsigned int textureId, unsigned int meshId, double x, double y,
+                            short angle);
+
             virtual ~PacketAddEntity() { };
 
-            virtual t_rawdata* serialize() const;
-            virtual void deserialize(t_rawdata* data);
+            virtual t_rawdata *serialize() const;
+            virtual void deserialize(t_rawdata *data);
             virtual size_t getPacketSize() const;
 
             unsigned int getEntityId() const;
             unsigned int getTextureId() const;
             unsigned int getMeshId() const;
-            int getX() const;
-            int getY() const;
+            double getX() const;
+            double getY() const;
             short getAngle() const;
 
         private:
             unsigned int entityId;
             unsigned int textureId;
             unsigned int meshId;
-            int x;
-            int y;
+            double x;
+            double y;
             short angle;
         };
     }
